+++ src/controllers/MenusController.php
@@ -9,6 +9,7 @@ use craft\web\Controller;
 use justinholt\freenav\assetbundles\FreeNavAsset;
 use justinholt\freenav\enums\Propagation;
 use justinholt\freenav\FreeNav;
+use justinholt\freenav\helpers\NodeHelper;
 use justinholt\freenav\models\Menu;
 use justinholt\freenav\models\MenuSiteSettings;
 use yii\web\NotFoundHttpException;
@@ -160,18 +161,7 @@ class MenusController extends Controller
         $nodeTypes = FreeNav::getInstance()->getNodeTypes()->getTypeOptions();
         $parentOptions = FreeNav::getInstance()->getNodes()->getParentOptions($menu);
 
-        // Build parent map from structure order: nodeId => parentId
-        $parentMap = [];
-        $stack = []; // stack of [id, level]
-        foreach ($nodes as $node) {
-            while (!empty($stack) && $stack[count($stack) - 1][1] >= $node->level) {
-                array_pop($stack);
-            }
-            if (!empty($stack)) {
-                $parentMap[$node->id] = $stack[count($stack) - 1][0];
-            }
-            $stack[] = [$node->id, $node->level];
-        }
+        $parentMap = NodeHelper::buildParentMap($nodes);
 
         return $this->renderTemplate('free-nav/menus/_build', [
             'menu' => $menu,